refactor common code from scripting tests; invalid tests fail by default
