+++ compiler/test/dotty/tools/scripting/BashScriptsTests.scala
@@ -2,29 +2,28 @@ package dotty
 package tools
 package scripting
 
-import java.io.File
-import java.nio.file.{Path, Paths, Files}
-import scala.sys.process._
-
 import org.junit.Test
 import org.junit.Assert.assertEquals
 
 import vulpix.TestConfiguration
 
-import dotty.tools.dotc.config.Properties._
+import ScriptTestEnv.*
 
 /** Verifies correct handling of command line arguments by `dist/bin/scala` and `dist/bin/scalac`.
  *   +. arguments following a script path must be treated as script arguments
  *   +. preserve script command line arguments.
+ *   +. prevent SCALA_OPTS in build environment from infecting tests, via 'SCALA_OPTS= ' prefix
+ *   +. test scripts must not throw execptions or exit with nonzero.
  */
 class BashScriptsTests:
   // classpath tests managed by scripting.ClasspathTests.scala
   def testFiles = scripts("/scripting")
+  lazy val argsfile = createArgsFile() // avoid problems caused by drive letter
 
   printf("osname[%s]\n", osname)
-  printf("using JAVA_HOME=%s\n", javaHome)
-  printf("using SCALA_HOME=%s\n", scalaHome)
-  printf("first 5 PATH entries:\n%s\n", pathEntries.take(5).mkString("\n"))
+  printf("using JAVA_HOME=%s\n", envJavaHome)
+  printf("using SCALA_HOME=%s\n", envScalaHome)
+  printf("first 5 PATH entries:\n%s\n", adjustedPathEntries.take(5).mkString("\n"))
   printf("scala path:  [%s]\n", scalaPath)
   printf("scalac path: [%s]\n", scalacPath)
 
@@ -44,9 +43,9 @@ class BashScriptsTests:
 
   /* verify `dist/bin/scalac` non-interference with command line args following script name */
   @Test def verifyScalacArgs =
-    val commandline = (Seq(scalacPath, "-script", showArgsScript) ++ testScriptArgs).mkString(" ")
+    val commandline = (Seq("SCALA_OPTS= ", scalacPath, "-script", showArgsScript) ++ testScriptArgs).mkString(" ")
     val (validTest, exitCode, stdout, stderr) = bashCommand(commandline)
-    if validTest then
+    if verifyValid(validTest) then
       var fail = false
       printf("\n")
       for (line, expect) <- stdout zip expectedOutput do
@@ -59,33 +58,33 @@ class BashScriptsTests:
 
   /* verify `dist/bin/scala` with -J setting */
   @Test def verifyScJProperty =
-    val commandline = Seq(scalaPath, "-J-Dkey=World", testFiles.find(_.getName == "envtest.sc").get.absPath).mkString(" ")
+    val commandline = Seq("SCALA_OPTS= ", scalaPath, "-J-Dkey=World", testFiles.find(_.getName == "envtest.sc").get.absPath).mkString(" ")
     val (validTest, exitCode, stdout, stderr) = bashCommand(commandline)
     assertEquals(stdout.mkString("/n"), "Hello World")
 
   /* verify `dist/bin/scala` with -J setting */
   @Test def verifyScalaJProperty =
-    val commandline = Seq(scalaPath, "-J-Dkey=World3", testFiles.find(_.getName == "envtest.scala").get.absPath).mkString(" ")
+    val commandline = Seq("SCALA_OPTS= ", scalaPath, "-J-Dkey=World3", testFiles.find(_.getName == "envtest.scala").get.absPath).mkString(" ")
     val (validTest, exitCode, stdout, stderr) = bashCommand(commandline)
     assertEquals(stdout.mkString("/n"), "Hello World3")
 
   /* verify `dist/bin/scala` with -D setting */
   @Test def verifyScDProperty =
-    val commandline = Seq(scalaPath, "-Dkey=World3", testFiles.find(_.getName == "envtest.sc").get.absPath).mkString(" ")
+    val commandline = Seq("SCALA_OPTS= ", scalaPath, "-Dkey=World3", testFiles.find(_.getName == "envtest.sc").get.absPath).mkString(" ")
     val (validTest, exitCode, stdout, stderr) = bashCommand(commandline)
     assertEquals(stdout.mkString("/n"), "Hello World3")
 
   /* verify `dist/bin/scala` with -D setting */
   @Test def verifyScalaDProperty =
-    val commandline = Seq(scalaPath, "-Dkey=World4", testFiles.find(_.getName == "envtest.scala").get.absPath).mkString(" ")
+    val commandline = Seq("SCALA_OPTS= ", scalaPath, "-Dkey=World4", testFiles.find(_.getName == "envtest.scala").get.absPath).mkString(" ")
     val (validTest, exitCode, stdout, stderr) = bashCommand(commandline)
     assertEquals(stdout.mkString("/n"), "Hello World4")
 
   /* verify `dist/bin/scala` with -D setting */
   @Test def saveAndRunWithDProperty =
-    val commandline = Seq(scalaPath, "-save", testFiles.find(_.getName == "envtest.scala").get.absPath).mkString(" ")
+    val commandline = Seq("SCALA_OPTS= ", scalaPath, "-save", testFiles.find(_.getName == "envtest.scala").get.absPath).mkString(" ")
     val (_, _, _, _) = bashCommand(commandline)
-    val commandline2 = Seq(scalaPath, "-Dkey=World5", testFiles.find(_.getName == "envtest.jar").get.absPath).mkString(" ")
+    val commandline2 = Seq("SCALA_OPTS= ", scalaPath, "-Dkey=World5", testFiles.find(_.getName == "envtest.jar").get.absPath).mkString(" ")
     val (validTest, exitCode, stdout, stderr) = bashCommand(commandline2)
     assertEquals(stdout.mkString("/n"), "Hello World5")
 
@@ -93,7 +92,7 @@ class BashScriptsTests:
   @Test def verifyScalaArgs =
     val commandline = (Seq("SCALA_OPTS= ", scalaPath, showArgsScript) ++ testScriptArgs).mkString(" ")
     val (validTest, exitCode, stdout, stderr) = bashCommand(commandline)
-    if validTest then
+    if verifyValid(validTest) then
       var fail = false
       printf("\n")
       var mismatches = List.empty[(String, String)]
@@ -115,7 +114,7 @@ class BashScriptsTests:
     printf("===> verify valid system property script.path is reported by script [%s]\n", scriptFile.getName)
     printf("calling scriptFile: %s\n", scriptFile)
     val (validTest, exitCode, stdout, stderr) = bashCommand(scriptFile.absPath)
-    if validTest then
+    if verifyValid(validTest) then
       stdout.foreach { printf("stdout: [%s]\n", _) }
       stderr.foreach { printf("stderr: [%s]\n", _) }
       val valid = stdout.exists { _.endsWith(expected) }
@@ -131,7 +130,7 @@ class BashScriptsTests:
     val envPairs = List(("SCALA_OPTS", s"@$argsfile"))
     val (validTest, exitCode, stdout, stderr) = bashCommand(scriptFile.absPath, envPairs)
     printf("stdout: %s\n", stdout.mkString("\n","\n",""))
-    if validTest then
+    if verifyValid(validTest) then
       val expected = s"${workingDirectory.norm}"
       val output = stdout.find( _.trim.startsWith("cwd") ).getOrElse("").dropWhile(_!=' ').trim
       printf("output  [%s]\n", output)
@@ -140,104 +139,3 @@ class BashScriptsTests:
       if valid then printf(s"\n===> success: classpath begins with %s, as reported by [%s]\n", workingDirectory, scriptFile.getName)
       assert(valid, s"script ${scriptFile.absPath} did not report valid java.class.path first entry")
 
-  def existingPath: String = envOrElse("PATH", "").norm
-  def adjustedPath = s"$javaHome/bin$psep$scalaHome/bin$psep$existingPath"
-  def pathEntries = adjustedPath.split(psep).toList
-
-  lazy val argsfile = createArgsFile() // avoid problems caused by drive letter
-  def createArgsFile(): String =
-    val utfCharset = java.nio.charset.StandardCharsets.UTF_8.name
-    val path = Files.createTempFile("scriptingTest", ".args")
-    val text = s"-classpath ${workingDirectory.absPath}"
-    Files.write(path, text.getBytes(utfCharset))
-    path.toFile.getAbsolutePath.norm
-
-  def fixHome(s: String): String =
-    s.startsWith("~") match {
-    case false => s
-    case true => s.replaceFirst("~", userHome)
-    }
-
-  extension(s: String) {
-    def toPath: Path = Paths.get(fixHome(s)) // .toAbsolutePath
-    def toFile: File = s.toPath.toFile
-    def absPath: String = s.toFile.absPath
-    def norm: String = s.replace('\\', '/') // bash expects forward slash
-    def isFile: Boolean = s.toFile.isFile
-    def exists: Boolean = s.toPath.toFile.exists
-    def name: String = s.toFile.getName
-    def dropExtension: String = s.reverse.dropWhile(_ != '.').drop(1).reverse
-    def parent(up: Int): String = s.norm.split("/").reverse.drop(up).reverse.mkString("/")
-  }
-
-  extension(p: Path) {
-    def listFiles: Seq[File] = p.toFile.listFiles.toList
-    def norm: String = p.normalize.toString.replace('\\', '/')
-    def name: String = p.toFile.getName
-  }
-
-  extension(f: File) {
-    def name = f.getName
-    def norm: String = f.toPath.normalize.norm
-    def absPath: String = f.getAbsolutePath.norm
-  }
-
-  lazy val psep: String = propOrElse("path.separator", "")
-  lazy val osname = propOrElse("os.name", "").toLowerCase
-
-  lazy val scalacPath = s"$workingDirectory/dist/target/pack/bin/scalac".norm
-  lazy val scalaPath = s"$workingDirectory/dist/target/pack/bin/scala".norm
-
-  // use optional working directory TEST_CWD, if defined
-  lazy val workingDirectory: String = envOrElse("TEST_CWD", userDir)
-
-  // use optional TEST_BASH if defined, otherwise, bash must be in PATH
-  lazy val bashExe: String = envOrElse("TEST_BASH", whichBash)
-
-  // test env SCALA_HOME is:
-  //    dist/target/pack, if present
-  //    else, SCALA_HOME if defined
-  //    else, not defined
-  lazy val scalaHome =
-    if scalacPath.isFile then scalacPath.replaceAll("/bin/scalac", "")
-    else envOrElse("SCALA_HOME", "").norm
-
-  lazy val javaHome = whichJava.parent(2)
-
-  lazy val testEnvPairs = List(
-    ("JAVA_HOME", javaHome),
-    ("SCALA_HOME", scalaHome),
-    ("PATH", adjustedPath),
-  ).filter { case (name, valu) => valu.nonEmpty }
-
-  lazy val whichBash: String = whichExe("bash")
-  lazy val whichJava: String = whichExe("java")
-
-  def whichExe(basename: String): String = 
-    val exeName = if (osname.toLowerCase.startsWith("windows")) s"$basename.exe" else basename
-    which(exeName)
-
-  def bashCommand(cmdstr: String, additionalEnvPairs: List[(String, String)] = Nil): (Boolean, Int, Seq[String], Seq[String]) = {
-    var (stdout, stderr) = (List.empty[String], List.empty[String])
-    if bashExe.toFile.exists then
-      val cmd = Seq(bashExe, "-c", cmdstr)
-      val envPairs = testEnvPairs ++ additionalEnvPairs
-      val proc = Process(cmd, None, envPairs *)
-      val exitVal = proc ! ProcessLogger (
-        (out: String) => stdout ::= out,
-        (err: String) => stderr ::= err
-      )
-      val validTest = exitVal == 0 && ! stderr.exists(_.contains("Permission denied"))
-      if ! validTest then
-        printf("\nunable to execute script, return value is %d\n", exitVal)
-        stderr.foreach { System.err.printf("stderr [%s]\n", _) }
-      (validTest, exitVal, stdout.reverse, stderr.reverse)
-    else
-      (false, -1, Nil, Nil)
-  }
-
-  def execCmd(command: String, options: String *): Seq[String] =
-    val cmd = (command :: options.toList).toSeq
-    for {
-      line <- Process(cmd).lazyLines_!
-    } yield line
